balancer1: session table

+++ modules/balancer/state.h
@@ -25,7 +25,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 struct worker_info {
-	_Atomic uint32_t use_prev_gen; // atomic
+	_Atomic uint8_t use_prev_gen; // atomic
 	uint8_t __padding[63];	       // NOLINT
 	_Atomic uint32_t max_deadline_current_gen;
 	_Atomic uint32_t max_deadline_prev_gen;
